requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27

pandas>=2.0
openpyxl>=3.1

sqlalchemy>=2.0

python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7.4
bcrypt<5

python-multipart>=0.0.9